some code improvements
