+++ lib/runtime/AllocationTracking.cpp
@@ -9,7 +9,6 @@
 #include "CallbackInterface.h"
 #include "Runtime.h"
 #include "RuntimeData.h"
-#include "RuntimeUtils.h"
 #include "TypeResolution.h"
 #include "support/Logger.h"
 
@@ -34,6 +33,49 @@ using namespace btree;
 #define RUNTIME_GUARD_END typeart::typeart_rt_scope = false
 
 namespace typeart {
+
+namespace detail {
+template <class...>
+constexpr std::false_type always_false{};
+}  // namespace detail
+
+template <typename Enum>
+inline Enum operator|(Enum lhs, Enum rhs) {
+  if constexpr (std::is_enum_v<Enum> && (std::is_same_v<Enum, AllocState> || std::is_same_v<Enum, FreeState>)) {
+    using enum_t = typename std::underlying_type<Enum>::type;
+    return static_cast<Enum>(static_cast<enum_t>(lhs) | static_cast<enum_t>(rhs));
+  } else {
+    static_assert(detail::always_false<Enum>);
+  }
+}
+template <typename Enum>
+inline void operator|=(Enum& lhs, Enum rhs) {
+  if constexpr (std::is_enum_v<Enum> && (std::is_same_v<Enum, AllocState> || std::is_same_v<Enum, FreeState>)) {
+    lhs = lhs | rhs;
+  } else {
+    static_assert(detail::always_false<Enum>);
+  }
+}
+
+template <typename Enum>
+inline Enum operator&(Enum lhs, Enum rhs) {
+  if constexpr (std::is_enum_v<Enum> && std::is_same_v<Enum, AllocState>) {
+    using enum_t = typename std::underlying_type<Enum>::type;
+    return static_cast<Enum>(static_cast<enum_t>(lhs) & static_cast<enum_t>(rhs));
+  } else {
+    static_assert(detail::always_false<Enum>);
+  }
+}
+
+template <typename Enum>
+inline typename std::underlying_type<Enum>::type operator==(Enum lhs, Enum rhs) {
+  if constexpr (std::is_enum_v<Enum> && std::is_same_v<Enum, AllocState>) {
+    using enum_t = typename std::underlying_type<Enum>::type;
+    return static_cast<enum_t>(lhs) & static_cast<enum_t>(rhs);
+  } else {
+    static_assert(detail::always_false<Enum>);
+  }
+}
 
 using namespace debug;
