--- lib/runtime/RuntimeUtils.cpp
@@ -1,9 +1,0 @@
-//
-// Created by sebastian on 08.01.21.
-//
-
-#include "RuntimeUtils.h"
-
-#include "TypeResolution.h"
-
-namespace typeart {}
