--- lib/runtime/RuntimeUtils.h
@@ -1,16 +1,0 @@
-//
-// Created by sebastian on 08.01.21.
-//
-
-#ifndef TYPEART_RUNTIMEUTILS_H
-#define TYPEART_RUNTIMEUTILS_H
-
-#include "RuntimeData.h"
-#include "RuntimeInterface.h"
-#include "support/Logger.h"
-
-#include <string>
-
-namespace typeart {}
-
-#endif  // TYPEART_RUNTIMEUTILS_H
